优化 spring-servlet.xml 的interceptors 配置

+++ superid-web-app/src/main/java/cn/superid/webapp/service/impl/UserService.java
@@ -71,7 +71,7 @@ public class UserService implements IUserService {
 
     @Override
     public boolean belong(long roleId) {
-       return RoleCache.dao.findFieldByKey(roleId,"userId",Long.class)==currentUserId();
+       return (Long)RoleCache.dao.findFieldByKey(roleId,"userId",Long.class)==currentUserId();
     }
